Location & Weather API merged at viewmodel

+++ app/src/main/java/com/ray/weather/data/remote/model/CurrentWeatherNetworkResponse.kt
@@ -8,7 +8,7 @@ data class CurrentWeatherNetworkResponse(
     @SerialName("temperature") val temperature: Double,
     @SerialName("windspeed") val windSpeed: Double,
     @SerialName("winddirection") val windDirection: Double,
-    @SerialName("weathercode") val weatherCode: Double,
+    @SerialName("weathercode") val weatherCode: Int,
     @SerialName("is_day") val isDay: Int,
     @SerialName("time") val time: String
 )
